Add files via upload
This issue content is generated by AI. Read it with caution.
I've recently developed a program for computing factorials using the difference method. At present, it's uncertain whether there are flaws in it. The program consists of four functions: factorial1, factorial2, factorial3, and factorial4. factorial1 is a conventional factorial - calculating function, merely serving as a comparison. factorial2, factorial3, and factorial4 represent novel approaches to calculating factorials grounded in the difference method. I earnestly invite everyone to explore this new algorithm together, assist in examining whether there are issues with the program, and optimize the algorithm.
Special instruction: This program needs to call the GMP library. Besides, I have another repository where Python - coded files are stored, and this code doesn't require the invocation of any libraries. It is hereby stated that in case the operation of this program leads to computer damage or file corruption, the operator and tester shall bear the responsibility.
Moreover, as this is my first time publishing on GitHub, I'm unsure which license to select. If the code is to be used in an open - source manner, I hope to be informed of the appropriate license.